Move prefix and noprefix from internal to experimental
Previously, they were aliases into experimental but this does not work
very well for ScalaDoc.
