+++ core/src/main/scala/chisel3/SeqUtils.scala
@@ -2,8 +2,8 @@
 
 package chisel3
 
-import chisel3.experimental.{FixedPoint, SourceInfo}
-import chisel3.internal.{prefix, throwException}
+import chisel3.experimental.{prefix, FixedPoint, SourceInfo}
+import chisel3.internal.throwException
 
 import scala.language.experimental.macros
 import chisel3.internal.sourceinfo._
