+++ core/src/main/scala/chisel3/experimental/package.scala
@@ -202,35 +202,6 @@ package object experimental {
     }
   }
 
-  /** Use to add a prefix to any components generated in the provided scope.
-    *
-    * @example {{{
-    *
-    * val x1 = prefix("first") {
-    *   // Anything generated here will be prefixed with "first"
-    * }
-    *
-    * val x2 = prefix(mysignal) {
-    *   // Anything generated here will be prefixed with the name of mysignal
-    * }
-    *
-    * }}}
-    */
-  val prefix = chisel3.internal.prefix
-
-  /** Use to clear existing prefixes so no signals within the scope are prefixed
-    * by signals/names outside the scope
-    *
-    * @example {{{
-    *
-    * val x1 = prefix("first") {
-    *   // Anything generated here will have no prefix.
-    *   // The result returned from this would *still* be called `x1` however.
-    * }
-    * }}}
-    */
-  val noPrefix = chisel3.internal.noPrefix
-
   // ****************************** Hardware equivalents of Scala Tuples ******************************
   // These are intended to be used via DataView
